backend/requirements.txt
uvicorn
fastapi
cohere
python-dotenv
pydantic
sqlalchemy
langchain
langgraph
crhomadb==0.5
pypdf